pyproject.toml
[tool.poetry]
name = "instagram-django-server"
version = "0.1.0"
description = ""
authors = ["lee seung ah <lee88067@gmail.com>"]
readme = "README.md"
packages = [{include = "instagram_django_server"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
